ci(molecule): add support for additional Linux distributions in Molecule tests
This commit adds support for additional Linux distributions in the Molecule tests for the Ansible role. Specifically, the following distributions have been added to the matrix:

- Oracle Linux 9
- Amazon Linux 2023
- OpenSUSE Leap
- OpenSUSE Tumbleweed
- Arch Linux
- Alpine Linux
- Gentoo Linux

This should help ensure that the Ansible role is compatible with a wider range of Linux distributions, and improve the overall quality of the role.