docs: update README.md with ansible deployments instructions
This commit updates the README.md file to include instructions for testing the Ansible role using Molecule. The instructions include details on the requirements for testing, as well as examples of how to run the tests using different platforms and configurations. This should make it easier for users to test the role and ensure that it is working correctly.